Purge only on production

+++ webapp/ruby/app.rb
@@ -150,8 +150,10 @@ class Isucon2App < Sinatra::Base
         "SELECT ticket_id FROM variation WHERE id = #{ mysql.escape(params[:variation_id]) } LIMIT 1",
       ).first['ticket_id']
 
-      purge_cache('http://127.0.0.1/')
-      purge_cache("http://127.0.0.1/ticket/#{ticket_id}")
+      if production?
+        purge_cache('http://127.0.0.1/')
+        purge_cache("http://127.0.0.1/ticket/#{ticket_id}")
+      end
 
       slim :complete, :locals => { :seat_id => seat_id, :member_id => params[:member_id] }
     else
